sdm-modbus-zeromq/rootfs/requirements.txt
sdm_modbus_zmq==0.0.1
paho-mqtt==1.6.1